Add script for automated 'pu' branch creation
Automated creation of a 'proposed-updates' branch speeds up the
gitworkflow.

The script follows the gitworkflow rules to create a 'proposed-updated'
branch. Topic branches which are already merged into 'next' will be
merged into 'pu'. Followed by topic branches which are currently not
integrated in 'next'.
Branches which are merged in 'master' or have a leading 'wip' will be
ignored.

On merge conflicts the script will exit. The user can then fix the
merge conflict. git-rerere will help to reuse the solution on the next
run of 'create-pu'.